requirements.txt
boto==2.49.0
click==7.1.2
DateTime==4.3
dnspython==2.1.0
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.0.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
passlib==1.7.4
pymongo==3.11.3
python-decouple==3.4
pytz==2021.1
SQLAlchemy==1.4.3
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
zope.interface==5.3.0
